Update test functions
